requirements.txt
qbittorrent-api==2025.5.0
requests==2.31.0
python-dotenv==1.0.0
telebot==0.0.5
colorama==0.4.6
tzdata==2025.1